cloud_functions/file_search_api/requirements.txt
functions-framework==3.*
google-genai>=1.0.0
flask==3.*
requests==2.*
